#8 Create tests for math library by Interface IMath
- Added tests: sum, sub, div, mult and fact
- Changed IMath.Fact argument type to double, becouse Calculator work
only with double numbers. IMath.Fact must check passed value by itself.

+++ src/Calculator/Math/IMath.cs
@@ -51,7 +51,7 @@ namespace Disassembler.Calculator.Math
 		/// </summary>
 		/// <param name="num">Factoriated number.</param>
 		/// <returns>Factorial from the number.</returns>
-		ulong Fact(ushort num);
+		ulong Fact(double num);
 
 		/// <summary>
 		///     Exponentiation with general exponent.
